requirements.txt
langchain
langchain-openai
langchain-community
openai
faiss-cpu
numpy
python-dotenv
streamlit
tavily-python